requirements.txt
aiohttp==2.2.5
async-timeout==1.3.0
asyncio==3.4.3
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
html5lib==0.999999999
idna==2.5
multidict==3.1.3
pymongo==3.5.0
python-engineio==1.7.0
python-socketio==1.8.0
requests==2.18.2
selenium==3.4.3
six==1.10.0
urllib3==1.22
webencodings==0.5.1
yarl==0.12.0
